DsBot/requirements.txt
disnake == 2.9.2
python-dotenv == 1.0.1